+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add dynamic object transformer
+- Add disjoint key transformer
 
 ### Changed
 - Increase length penalty for primary key detection